+++ boilerplate/.cache/07/fcb901c6fcf2fd2046889fcc4848b6.json
@@ -1,1 +1,1 @@
-{"id":"index.html","dependencies":[{"name":"./styles.scss","dynamic":true,"resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\styles.scss","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\index.html"},{"name":"./ts\\index.ts","dynamic":true,"resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\ts\\index.ts","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\index.html"}],"generated":{"html":"<!DOCTYPE html>\r\n<html lang=\"en\">\r\n<head>\r\n    <meta charset=\"UTF-8\">\r\n    <meta http-equiv=\"X-UA-Compatible\" content=\"IE=edge\">\r\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\r\n    <title>Interval App</title>\r\n    <link rel=\"stylesheet\" href=\"/238b3221445c00e4c29b1d93164d45a1.css\">\r\n</head>\r\n<body>\r\n    <div id=\"app\">\r\n\r\n    </div>\r\n    <script src=\"/2d03800ce419d2ac7934b2b9841fc46b.js\"></script>\r\n</body>\r\n</html>"},"sourceMaps":null,"error":null,"hash":"e6bb91c12e73bd3990ae56fb92cb3178","cacheData":{}}
+{"id":"index.html","dependencies":[{"name":"./styles.scss","dynamic":true,"resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\styles.scss","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\index.html"},{"name":"./img\\logo.svg","dynamic":true,"resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\img\\logo.svg","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\index.html"},{"name":"./ts\\index.ts","dynamic":true,"resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\ts\\index.ts","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\index.html"}],"generated":{"html":"<!DOCTYPE html>\r\n<html lang=\"en\">\r\n<head>\r\n    <meta charset=\"UTF-8\">\r\n    <meta http-equiv=\"X-UA-Compatible\" content=\"IE=edge\">\r\n    <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\r\n    <title>Interval App</title>\r\n    <link rel=\"stylesheet\" href=\"/238b3221445c00e4c29b1d93164d45a1.css\">\r\n</head>\r\n<body>\r\n    <div id=\"app\">\r\n        <section id=\"loading\" class=\"show\">\r\n            <a href=\"#\" class=\"btn\" data-set-view=\"set-timer\">\r\n                <img src=\"/b3f8a10f35814969f4c5fbfaf1d8a4c6.svg\" alt=\"logo\">\r\n                <p>For all your timing needs.</p>\r\n            </a>\r\n        </section>\r\n    </div>\r\n    <script src=\"/2d03800ce419d2ac7934b2b9841fc46b.js\"></script>\r\n</body>\r\n</html>"},"sourceMaps":null,"error":null,"hash":"8bc0205494901fcf46832e21d4f7ecff","cacheData":{}}
