+++ boilerplate/.cache/cd/bcf6766b1dbcc28264058a8fd6340a.json
@@ -1,0 +1,1 @@
+{"id":"ts/loading.ts","dependencies":[{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\tsconfig.json","includedInParent":true,"mtime":1621324332816},{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\package.json","includedInParent":true,"mtime":1621329912572}],"generated":{"js":"\"use strict\";\n\nfunction loading() {\n  var template = \"\\n    <section id=\\\"timer-digital class=\\\"show\\\">\\n        \\n    </section> \\n    \";\n}"},"sourceMaps":{"js":{"mappings":[{"source":"ts/loading.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":3,"column":0}},{"source":"ts/loading.ts","name":null,"original":{"line":1,"column":9},"generated":{"line":3,"column":9}},{"source":"ts/loading.ts","name":null,"original":{"line":1,"column":0},"generated":{"line":3,"column":16}},{"source":"ts/loading.ts","name":null,"original":{"line":1,"column":16},"generated":{"line":3,"column":19}},{"source":"ts/loading.ts","name":null,"original":{"line":3,"column":4},"generated":{"line":4,"column":0}},{"source":"ts/loading.ts","name":null,"original":{"line":3,"column":8},"generated":{"line":4,"column":6}},{"source":"ts/loading.ts","name":null,"original":{"line":3,"column":16},"generated":{"line":4,"column":14}},{"source":"ts/loading.ts","name":null,"original":{"line":3,"column":27},"generated":{"line":4,"column":17}},{"source":"ts/loading.ts","name":null,"original":{"line":3,"column":4},"generated":{"line":4,"column":101}},{"source":"ts/loading.ts","name":null,"original":{"line":8,"column":1},"generated":{"line":5,"column":0}}],"sources":{"ts/loading.ts":"function loading(){\r\n\r\n    let template: string = `\r\n    <section id=\"timer-digital class=\"show\">\r\n        \r\n    </section> \r\n    `;\r\n}"},"lineCount":null}},"error":null,"hash":"0ba20a26ebf4a32864f34ee413d9d6b7","cacheData":{"env":{}}}
