+++ boilerplate/.cache/e6/8cead551762b65bfbdeabef5810260.json
@@ -1,0 +1,1 @@
+{"id":"img/logo.svg","dependencies":[{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\package.json","includedInParent":true,"mtime":1621329912572}],"generated":{"js":"module.exports = \"/b3f8a10f35814969f4c5fbfaf1d8a4c6.svg\";"},"sourceMaps":null,"error":null,"hash":"8bcf6ac83d8cae88812f4b9aac55a260","cacheData":{"env":{}}}
