+++ boilerplate/.cache/fb/4016c1fc58af5bef215ece8851e8ec.json
@@ -1,1 +1,1 @@
-{"id":"styles.scss","dependencies":[{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\styles.scss","includedInParent":true,"mtime":1621339330833},{"name":"_css_loader","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\styles.scss","resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\node_modules\\parcel-bundler\\src\\builtins\\css-loader.js"},{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\package.json","includedInParent":true,"mtime":1621329912572}],"generated":{"css":"body {\n  background: black;\n}\nbody h1 {\n  color: white;\n}\nbody #timer-digital {\n  border: 1px solid white;\n}","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"sourceMaps":{"css":{"mappings":[{"source":"styles.scss","name":null,"original":{"line":1,"column":0},"generated":{"line":1,"column":0}},{"source":"styles.scss","name":null,"original":{"line":2,"column":4},"generated":{"line":2,"column":2}},{"source":"styles.scss","name":null,"original":{"line":4,"column":4},"generated":{"line":4,"column":0}},{"source":"styles.scss","name":null,"original":{"line":5,"column":8},"generated":{"line":5,"column":2}},{"source":"styles.scss","name":null,"original":{"line":8,"column":4},"generated":{"line":7,"column":0}},{"source":"styles.scss","name":null,"original":{"line":9,"column":8},"generated":{"line":8,"column":2}}],"sources":{"styles.scss":"body {\r\n    background: black;\r\n\r\n    h1 {\r\n        color: white;\r\n    }\r\n\r\n    #timer-digital {\r\n        border: 1px solid white;\r\n    }\r\n}"},"lineCount":null}},"error":null,"hash":"6c2f058945f8d6465c0a7234390e43ae","cacheData":{"env":{}}}
+{"id":"styles.scss","dependencies":[{"name":"_css_loader","parent":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\src\\styles.scss","resolved":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\node_modules\\parcel-bundler\\src\\builtins\\css-loader.js"},{"name":"C:\\Users\\johnh\\OneDrive\\Skrivbord\\MightyEagle\\Hämtade från Git\\dundergruppen\\boilerplate\\package.json","includedInParent":true,"mtime":1621329912572}],"generated":{"css":"","js":"var reloadCSS = require('_css_loader');\n\nmodule.hot.dispose(reloadCSS);\nmodule.hot.accept(reloadCSS);"},"sourceMaps":{"css":{"mappings":[{"source":"styles.scss","original":{"line":1,"column":0},"generated":{"line":1,"column":0}}],"sources":{"styles.scss":""},"lineCount":1}},"error":null,"hash":"e85a34b625b8e8363310bee3f73a0792","cacheData":{"env":{}}}
