+++ boilerplate/src/index.html
@@ -9,7 +9,12 @@
 </head>
 <body>
     <div id="app">
-
+        <section id="loading" class="show">
+            <a href="#" class="btn" data-set-view="set-timer">
+                <img src="./img/logo.svg" alt="logo">
+                <p>For all your timing needs.</p>
+            </a>
+        </section>
     </div>
     <script src="./ts/index.ts"></script>
 </body>
