--- boilerplate/src/styles.scss
@@ -1,11 +1,0 @@
-body {
-    background: black;
-
-    h1 {
-        color: white;
-    }
-
-    #timer-digital {
-        border: 1px solid white;
-    }
-}
