--- boilerplate/src/ts/abortTimer.ts
@@ -1,10 +1,0 @@
-function render() {
-
-    let template:string = `
-    <section id="abort-section">
-        <button id="abort-btn">Abort timer</button>
-    </section>
-    `
-}
-
-export { render };
