+++ boilerplate/src/ts/loading.ts
@@ -1,0 +1,8 @@
+function loading(){
+
+    let template: string = `
+    <section id="timer-digital class="show">
+        
+    </section> 
+    `;
+}
